+++ Common.Test/LoggerTest.cs
@@ -11,10 +11,10 @@ public class LoggerTest
     public void ShouldLogToFIle()
     {
         //Arrange
-        var logger = new Logger("Test");
+        var logger = new Logger("Test", "Test");
         //Act
         logger.Log("Test", "Test", LogType.INFO);
-        var exists = System.IO.File.Exists($"./bin/pirateTest/logs/{logger.logName}.log");
+        var exists = System.IO.File.Exists($"./bin/pirateTest/logs/Test.log");
 
         //Assert
         Assert.True(exists);