+++ Common/ILogger.cs
@@ -4,7 +4,6 @@ namespace Common;
 
 public interface ILogger
 {
-    string logName { get; set; }
 
     void Log(string message, string orginFile, LogType logType);
 }