+++ Interpreter/Interpreters/ComparisonOperationNodeInterpreter.cs
@@ -2,6 +2,7 @@ using Interpreter.Values;
 using Parser.Node.Interfaces;
 using Lexer.Enums;
 using Common;
+using Common.Errors;
 
 namespace Interpreter.Interpreters;
 
@@ -9,12 +10,12 @@ public class ComparisonOperationNodeInterpreter : BaseInterpreter
 {
     public IOperationNode Node { get; set; }
     private InterpreterFactory interpreterFactory;
-    public ILogger Logger { get; set; }
-    public ComparisonOperationNodeInterpreter(INode node, InterpreterFactory InterpreterFactory, ILogger logger)
+    public ComparisonOperationNodeInterpreter(INode node, InterpreterFactory InterpreterFactory, ILogger logger) : base(logger)
     {
-        Node = node as IOperationNode;
+        if (node is not IOperationNode) throw new TypeConversionException(node.GetType(), typeof(IOperationNode));
+        Node = (IOperationNode)node;
+        
         interpreterFactory = InterpreterFactory;
-        Logger = logger;
         Logger.Log($"Created {this.GetType().Name} : \"{Node.ToString()}\"", this.GetType().Name, Common.Enum.LogType.INFO);
     }