+++ Interpreter/Interpreters/VariableAssignNodeInterpreter.cs
@@ -3,6 +3,7 @@ using Parser.Node.Interfaces;
 using Common;
 using Interpreter.Values;
 using Parser.Node;
+using Common.Errors;
 
 namespace Interpreter.Interpreters;
 
@@ -10,25 +11,31 @@ public class VariableAssignNodeInterpreter : BaseInterpreter
 {
     public VariableAssignNode Node { get; set; }
     private InterpreterFactory interpreterFactory;
-    public ILogger Logger { get; set; }
 
-    public VariableAssignNodeInterpreter(INode node, InterpreterFactory InterpreterFactory, ILogger logger)
+    public VariableAssignNodeInterpreter(INode node, InterpreterFactory InterpreterFactory, ILogger logger) : base(logger)
     {
-        Node = node as VariableAssignNode;
+        if (node is not VariableAssignNode)throw new TypeConversionException(node.GetType(), typeof(VariableAssignNode));
+
+        Node = (VariableAssignNode)node;
+
         interpreterFactory = InterpreterFactory;
-        Logger = logger;
         Logger.Log($"Created {this.GetType().Name} : \"{Node.ToString()}\"", this.GetType().Name, Common.Enum.LogType.INFO);
     }
 
     public override BaseValue VisitNode()
     {
-        var Identifier = Node.Identifier.Value.Value;
-        
+        if (Node.Identifier.Value.Value is not string)
+        {
+            if (Node.Identifier.Value.Value != null)
+            {
+                throw new TypeConversionException(Node.Identifier.Value.Value.GetType(), typeof(string));
+            }
+            throw new TypeConversionException(typeof(string));
+        }
+        var Identifier = (string)Node.Identifier.Value.Value;
+        SymbolTable.Instance(Logger).Set(Identifier, Node.Value);
 
-        SymbolTable.Instance(Logger).Set((string)Identifier, Node.Value);
-
-        var variable = new Variable((string)Identifier, Logger);
-
+        var variable = new Variable(Identifier, Logger);
         return variable;
     }
 }